# bin/itclub-reddit-api.swagger.json
{
  "swagger": "2.0",
  "paths": {
    "/": {
      "get": {
        "summary": "/",
        "consumes": [
          "application/json"
        ],
        "parameters": []
      },
      "delete": {
        "summary": "/",
        "consumes": [
          "application/json"
        ],
        "parameters": [],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Storage cleared!"
                }
              }
            }
          }
        }
      }
    },
    "/activeUser": {
      "put": {
        "summary": "/activeUser",
        "consumes": [
          "application/json"
        ],
        "parameters": [],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "1 docs updated."
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "/activeUser",
        "consumes": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string",
                  "example": "Zmotey"
                },
                "password": {
                  "type": "string",
                  "example": "******"
                },
                "__type": {
                  "type": "string",
                  "example": "User"
                },
                "_id": {
                  "type": "string",
                  "example": "oXSYWIeKEPzs2Gdq"
                }
              }
            }
          },
          "304": {},
          "404": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "No active User found!"
                }
              }
            }
          }
        },
        "produces": [
          "application/json"
        ]
      },
      "delete": {
        "summary": "/activeUser",
        "consumes": [
          "application/json"
        ],
        "parameters": []
      }
    },
    "/activeTopic": {
      "put": {
        "summary": "/activeTopic",
        "consumes": [
          "application/json"
        ],
        "parameters": [],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "1 docs updated."
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "/activeTopic",
        "consumes": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "ALL"
                },
                "__type": {
                  "type": "string",
                  "example": "Topic"
                },
                "_id": {
                  "type": "string",
                  "example": "3yw5XTg63ORDDIbB"
                }
              }
            }
          },
          "304": {}
        },
        "produces": [
          "application/json"
        ]
      },
      "delete": {
        "summary": "/activeTopic",
        "consumes": [
          "application/json"
        ],
        "parameters": []
      }
    },
    "/users": {
      "get": {
        "summary": "/users",
        "consumes": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "__type": {
                    "type": "string"
                  },
                  "_id": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "304": {}
        },
        "produces": [
          "application/json"
        ]
      },
      "post": {
        "summary": "/users",
        "consumes": [
          "application/json"
        ],
        "parameters": [],
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string",
                  "example": "Zmotey"
                },
                "password": {
                  "type": "string",
                  "example": "******"
                },
                "__type": {
                  "type": "string",
                  "example": "User"
                },
                "_id": {
                  "type": "string",
                  "example": "wMRDKg5ol8nc5b8j"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "/users",
        "consumes": [
          "application/json"
        ],
        "parameters": [],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Users storage cleared!"
                }
              }
            }
          }
        }
      }
    },
    "/users/{username}": {
      "get": {
        "summary": "/users/{username}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true
          }
        ]
      },
      "put": {
        "summary": "/users/{username}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true
          }
        ]
      },
      "delete": {
        "summary": "/users/{username}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true
          }
        ]
      }
    },
    "/topics": {
      "get": {
        "summary": "/topics",
        "consumes": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "schema": {
              "type": "array",
              "items": {}
            }
          },
          "304": {}
        },
        "produces": [
          "application/json"
        ]
      },
      "post": {
        "summary": "/topics",
        "consumes": [
          "application/json"
        ],
        "parameters": [],
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "ALL"
                },
                "__type": {
                  "type": "string",
                  "example": "Topic"
                },
                "_id": {
                  "type": "string",
                  "example": "UUDbmCVvM7mGj2HS"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "/topics",
        "consumes": [
          "application/json"
        ],
        "parameters": [],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Topics storage cleared!"
                }
              }
            }
          }
        }
      }
    },
    "/topics/{topicname}": {
      "get": {
        "summary": "/topics/{topicname}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "topicname",
            "in": "path",
            "required": true
          }
        ]
      },
      "put": {
        "summary": "/topics/{topicname}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "topicname",
            "in": "path",
            "required": true
          }
        ]
      },
      "delete": {
        "summary": "/topics/{topicname}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "topicname",
            "in": "path",
            "required": true
          }
        ]
      }
    },
    "/posts": {
      "get": {
        "summary": "/posts",
        "consumes": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "schema": {
              "type": "array",
              "items": {}
            }
          },
          "304": {}
        },
        "produces": [
          "application/json"
        ]
      },
      "post": {
        "summary": "/posts",
        "consumes": [
          "application/json"
        ],
        "parameters": [],
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "schema": {
              "type": "object",
              "properties": {
                "_id": {
                  "type": "number",
                  "example": 1543524751086
                },
                "topic": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "example": "ALL"
                    },
                    "__type": {
                      "type": "string",
                      "example": "Topic"
                    }
                  }
                },
                "owner": {
                  "type": "object",
                  "properties": {
                    "username": {
                      "type": "string",
                      "example": "Zmotey"
                    },
                    "password": {
                      "type": "string",
                      "example": "******"
                    },
                    "__type": {
                      "type": "string",
                      "example": "User"
                    }
                  }
                },
                "title": {
                  "type": "string",
                  "example": "Header 5"
                },
                "image": {
                  "type": "string",
                  "example": "https://picsum.photos/800/400?image=5"
                },
                "message": {
                  "type": "string",
                  "example": "Kevin leberkas sed doner, esse magna kielbasa.  Landjaeger sed elit, deserunt chuck non leberkas esse dolore dolor pork belly short ribs spare ribs laboris.  Leberkas hamburger andouille, shankle anim aliquip burgdoggen tenderloin tempor.  Elit reprehenderit turkey anim, irure boudin tail commodo excepteur bacon ribeye.  Labore irure in do, in adipisicing sunt consequat andouille pork loin magna pork chop hamburger."
                },
                "comments": {
                  "type": "array",
                  "items": {},
                  "example": []
                },
                "upVotes": {
                  "type": "array",
                  "items": {},
                  "example": []
                },
                "downVotes": {
                  "type": "array",
                  "items": {},
                  "example": []
                },
                "__type": {
                  "type": "string",
                  "example": "Post"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "/posts",
        "consumes": [
          "application/json"
        ],
        "parameters": [],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "All Posts updated successfully."
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "/posts",
        "consumes": [
          "application/json"
        ],
        "parameters": [],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Posts storage cleared!"
                }
              }
            }
          }
        }
      }
    },
    "/posts/{postId}": {
      "get": {
        "summary": "/posts/{postId}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true
          }
        ]
      },
      "put": {
        "summary": "/posts/{postId}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true
          }
        ]
      },
      "delete": {
        "summary": "/posts/{postId}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true
          }
        ]
      }
    }
  },
  "info": {
    "description": "[Specification JSON](/api-spec)"
  }
}